Import packagecloud.io's GPG key to fix linux build. (#64)
The build started failing because an `apt-get install git-lfs` starting failing due to not having the source's, packagecloud..io's GPG key. This apt-get was not part of .travis.yml so

    1. The November updates to the Travis Xenial environment starting installing git-lfs,
    2. One of the packages we are installing was updated to have a dependency on git-lfs,
    3. or GPG started throwing an error instead of a warning when it couldn't verify a package. 

Since something was implicitly installing the package via apt-get, this fix also deletes the manual install in favour of an explicit apt-get.