Update vulkansdk to 1.2.176.1 and VK_KHR_portability_subset (#436)
* Update information about iOS app signing.

* Make compliant with VK_KHR_portability_subset.

This change looks MUCH larger than it really is. The main change
is in VulkanAppSDL.cpp and VulkanContext.h plus Texture.cpp is
modified to check if it can use swizzle before doing so. The vast
majority of the other changes are to streamline the vulkan.h &
vulkan.hpp includes to minimize the number of places we have to
prefix with VK_ENABLE_BETA_EXTENSIONS to access VK_KHR_portability_subset
declarations. The changes in VulkanSwapChain.cpp and vulkantextoverlay.hpp
are solely to workaround an overly strict validation rule.