Prioritize release tag when generating versions. Fixes #467 (#475)
* Prioritize release over rc over beta tags for versions.

Eliminate confusion caused by generating versions relative to rc or beta tags when there is is a newer release tag but no changes between, e.g., beta and release.

* Have mkversion cd to its location (repo root) before running.

Not running in the root during the Linux build is the only reason for the Linux version numbers reported in #467 but whatever caused mkversion to not be run in root appears to have been fixed. It is not possible to reproduce in current master prior to this PR.

* Try apt-get update in before_install to see if it fixes Linux build issue.

Unrelated fix to Linux CI builds.